feat(NT-312) updates ProductCardMain props and makes GA and GTM events implicit in the function
The idea here is that the sendGAEvent and sendGTM functions are only imported and called in the productcardmain component and the variable which is the event name is passed as a prop. This is because it is only the event name that is changing.
Also note that the the default export that you were removing was for the ProductCardMain. That way when you import it you have to import it with exactly the name you defined it as.